requirements.txt
openai
python-dotenv
streamlit
pandas
jsonlines
dspy-ai